refactor(solver): move mapped-keys re-entrant guard to a child module to respect the mapped.rs shard ratchet
The guard added 25 lines to evaluate_rules/mapped.rs (already at its
1982-line ratchet baseline); host it in mapped/keys_guard.rs instead and
keep mapped.rs at the baseline. No behavior change.

AgentName: StudioOpus
